% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedPlot.R
\name{combinedPlot}
\alias{combinedPlot}
\title{Plotting sequence preferences}
\usage{
combinedPlot(fc_results, i_cont, main = "")
}
\arguments{
\item{fc_results}{Output of "enrichmentMatrix"}

\item{i_cont}{Output of "iContent"}

\item{main}{Title of the plot}
}
\description{
Function plots a heatmap of sequence preferences 
and optionally a barplot with information content
}
\author{
Lukasz Jan Kielpinski
}
\seealso{
\code{\link{enrichmentMatrix}}

\code{\link{iContent}}
}

