% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentMatrix.R
\name{enrichmentMatrix}
\alias{enrichmentMatrix}
\title{Makes PWM from differential abundunce data}
\usage{
enrichmentMatrix(DEresults, enzyme_source, constr, smallestUnit = 1)
}
\arguments{
\item{DEresults}{Output of "GetKmerDE"}

\item{enzyme_source}{Enzyme of interest}

\item{constr}{Construct of interest}

\item{smallestUnit}{Length of a kmer (1 = single nucleotide; 2 = dinucleotide, ...)}
}
\value{
Matrix with log2FC associated with a given kmer at a given position for a given enzyme
}
\description{
The function extracts information for a selected enzyme, selected construct 
and selected kmer-lengthfrom a list of differential abundunce 
data generated by \code{\link{GetKmerDE}}. It creates a matrix, in which 
every row
corresponds to different kmer sequence (4^smallestUnit rows) and every column
corresponds to different starting position within a construct.
}
\author{
Lukasz Jan Kielpinski
}
\seealso{
\code{\link{GetKmerDE}}
}

