% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCleavageSite.R
\name{findCleavageSite}
\alias{findCleavageSite}
\title{Assigns preferred cleavage site of HIV-1 RNase H within a heptamer}
\usage{
findCleavageSite(fc_results, kmerLen = 7)
}
\arguments{
\item{fc_results}{Output of the "enrichmentMatrix" function with smallestUnit set to 7}

\item{kmerLen}{Length of the cleaved kmer (Currently the function works only for kmerLen = 7)}
}
\value{
cut_site Vector with numbers indicating after which nucleotide of each provided heptamer HIV-1 RNase H preferentially cleaves
}
\description{
Function assigns where within a given heptamer, the HIV-1 RNase H is expected to cleave.
Implementation works only with R7 construct, 
based on the observation described in the H-SPA publication.
}
\author{
Lukasz Jan Kielpinski
}
\seealso{
\code{\link{enrichmentMatrix}}
}

