% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetKmerDE.R
\name{GetKmerDE}
\alias{GetKmerDE}
\title{Calculates differential abundunce of kmers}
\usage{
GetKmerDE(all_kmers, samples)
}
\arguments{
\item{all_kmers}{List structure of counts of kmers at different position 
generated by \code{\link{importKmers}}}

\item{samples}{A table with the description of samples. Guides the function 
on what to compare to what.}
}
\value{
DEresults - Multi-level list structure. The first level are different 
enzymes, the second level are different constructs,
the third level is kmer length, 
the fourth level is kmer starting position within a construct, 
and the fifth level is a type of information for each kmer: 
(a) logFC - log2 fold change, 
(b) PValue - p-value that the abundunce has significantly changed
(c) logCPM - log2 counts-per-million  
(d) pi_value - pi-value, defined as -log10(p-value)*(fold change) (see Xiao Y et al. 2012)
which contain a vector of a values for different kmers (sorted alphabetically)
}
\description{
Function uses edgeR package to calculate differences 
between abundunces of kmers at different positions of the probed construct.
}
\author{
Lukasz Jan Kielpinski
}
\references{
Robinson, Mark D., Davis J. McCarthy, and Gordon K. Smyth. 2010. 
“edgeR: A Bioconductor Package for Differential Expression Analysis of 
Digital Gene Expression Data.” Bioinformatics (Oxford, England) 26 (1):
139–40. doi:10.1093/bioinformatics/btp616.

Xiao, Yufei, Tzu-Hung Hsiao, Uthra Suresh, Hung-I. Harry Chen, 
Xiaowu Wu, Steven E. Wolf, and Yidong Chen. 2014. “A Novel Significance 
Score for Gene Selection and Ranking.” Bioinformatics (Oxford, England) 
30 (6): 801–7. doi:10.1093/bioinformatics/btr671.
}
\seealso{
\code{\link{importKmers}}
}

