% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iContent.R
\name{iContent}
\alias{iContent}
\title{Calculate information content from the changes in position-specific nucleotide abundunce}
\usage{
iContent(DEresults, enzyme_source, constr)
}
\arguments{
\item{DEresults}{Output of "GetKmerDE"}

\item{enzyme_source}{Enzyme of interest}

\item{constr}{Construct of interest}
}
\value{
i_cont Vector with information content (Bits)
}
\description{
Function calculates information content of different nucleotide positions in a construct.
Based on a formula from Gorodkin et al. 1997. 
Note that the information content is calculated on the whole set, 
hence it is a function of a probing extent.
}
\author{
Lukasz Jan Kielpinski
}
\references{
Gorodkin, J., L. J. Heyer, S. Brunak, and G. D. Stormo. 1997. 
“Displaying the Information Contents of Structural RNA Alignments: 
The Structure Logos.” Computer Applications in the Biosciences: 
CABIOS 13 (6): 583–86.
}

