% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importKmers.R
\name{importKmers}
\alias{importKmers}
\title{Importing counts of kmers from structure of directories}
\usage{
importKmers(dataDir = ".", indices, constructs, maxKmerLen = 8)
}
\arguments{
\item{dataDir}{Path to the main directory with the processed data (do not include "/" at the end)}

\item{indices}{Indexes to be imported. If not provided, all dirs from the dataDir will be used.}

\item{constructs}{Names of the constructs to be imported. 
If not provided, all files with ".txt.gz" at the proper location extension will be imported.}

\item{maxKmerLen}{The longest kmer length that will be imported (default = 8)}
}
\value{
all_kmers Multi-level list structure. The first level are different constructs,
the second level is sequencing index of the sample, 
the third level is kmer length, 
which contains a matrix of counts of different kmers at different positions of the constructs
}
\description{
Function imports data generated from FASTQ files by bash scripts
and creates a standardized R list.
}
\author{
Lukasz Jan Kielpinski
}

