% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localDinucleotideSampling.R
\name{localDinucleotideSampling}
\alias{localDinucleotideSampling}
\title{Random sampling of the nucleotide sequence with statistical preservation of 
the local dinucleotide content}
\usage{
localDinucleotideSampling(char, window_size)
}
\arguments{
\item{char}{Original sequence to be sampled}

\item{window_size}{Width of the window from which dinucleotide content should be calculated. 
Window is centered around sampled nucleotide.}
}
\value{
Sampled sequence
}
\description{
Function generates a randomized version of the input DNA sequence, which is 
of the same length as the input sequence.
For position 'i' the occurences of dinucleotides are counted in the input sequence within a window 
from 'i - window_size/2' to 'i + window_size/2' and a new dinucleotide is selected according to:
(a) Narrow the set of dinucleotides to those which start from the second nucleotide of a previous dinucleotide.
(b) Randomly selected a new dinucleotide from the set defined in (a) with probabilities proportional to number 
of occurences of dinucleotides in the window. If occurences == 0, then select one by random. 
For the first dinucleotide: skip point (a).
}
\author{
Lukasz Jan Kielpinski
}
\seealso{
\code{\link{globalDinucleotideSampling}}
}

