% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hbp.R
\name{make_hbp}
\alias{make_hbp}
\title{Make Heptamer Based Predictor for HIV-1 RNase H}
\usage{
make_hbp(DEresults, enzyme_source = "HIV", constr)
}
\arguments{
\item{DEresults}{Output of "GetKmerDE"}

\item{enzyme_source}{Enzyme of interest. Function intended to be used only with HIV-1 RNase H.}

\item{constr}{Construct of interest. Function intended to be used only with R7.}
}
\value{
hbp Heptamer Based Predictor (List of kmers_seq, where_cut and logFC_cut)
}
\description{
Function makes the hbp (Heptamer Based Predictor) from differential abundances
Only for HIV-1 RNase H
}
\author{
Lukasz Jan Kielpinski
}

