% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAllKmers.R
\name{makeAllKmers}
\alias{makeAllKmers}
\title{Generate all kmers of a given length}
\usage{
makeAllKmers(smallestUnit, nts = c("A", "C", "G", "T"))
}
\arguments{
\item{smallestUnit}{Lenght of kmers to be generated}

\item{nts}{Nucleotides to use (default = c("A", "C", "G", "T"))}
}
\value{
Sorted vector of kmers (character)
}
\description{
Function generates a character vector with all possible DNA kmers of a given length (sorted)
}
\author{
Lukasz Jan Kielpinski
}

