% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNamedKmers.R
\name{makeNamedKmers}
\alias{makeNamedKmers}
\title{Generate variables with all kmers of a given length}
\usage{
makeNamedKmers(whichToMake)
}
\arguments{
\item{whichToMake}{Variables for which lengths should be created.}
}
\value{
Creates a variable in a Global Environment (unless variable of the same name exists in a parent environment which is not global environment)
}
\description{
Function generates variables with character vectors with all possible DNA kmers of given lengths (sorted)
Creates variables names dimers, trimers, tetramers, pentamers, hexamers, heptamers, octamers and nonamers
for 2 to 9 mers.
}
\author{
Lukasz Jan Kielpinski
}

