% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeViolinPlot.R
\name{makeViolinPlot}
\alias{makeViolinPlot}
\title{Make Violin Plot comparing counts in a reference set to counts in background sets}
\usage{
makeViolinPlot(countsInBinsList)
}
\arguments{
\item{countsInBinsList}{Result of \code{\link{summarizeBins}}}
}
\value{
Violin plot
}
\description{
Function uses ggplot2 to plot "violins" of background distribution (e.g. sampled content)
and prints rombi at the counts of a reference set.
}
\author{
Lukasz Jan Kielpinski
}

