% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotArea.R
\name{plotArea}
\alias{plotArea}
\title{Plot HIV-1 RNase H cleavage scores together with a sequence}
\usage{
plotArea(rnaSequence, CleavageScores, hbp, startPos, endPos,
  xlab = "Bond number", ...)
}
\arguments{
\item{rnaSequence}{Sequence of the RNA strand (T will be converted to U for display)}

\item{CleavageScores}{Cleavage scores to plot}

\item{hbp}{If CleavageScores are not provided, hbp and rnaSequence will be used to call \code{\link{predictCleavages}} and calcaulte the scores. Look there for more info.}

\item{startPos}{From which position to plot}

\item{endPos}{To which position to plot}

\item{xlab}{X-axis label}

\item{...}{Further parameters to \code{\link{plot}}}
}
\value{
Plot
}
\description{
Function plots cleavability score (based on a log2-fold-change of a heptamer in the probing experiment) 
of bonds within a provided sequence. It takes as input predicted cleavage scores or 
alternatively hbp list with which it runs the prediction internally.
}
\author{
Lukasz Jan Kielpinski
}
\seealso{
\code{\link{predictCleavages}}
}

