% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{plotCorrelation}
\alias{plotCorrelation}
\title{Simple convenience function for plotting observed vs predicted log-fold-changes}
\usage{
plotCorrelation(observed, predicted, col = "#00000010", ...)
}
\arguments{
\item{observed}{observed logFC}

\item{predicted}{predicted logFC}
}
\description{
Simple convenience function for plotting observed vs predicted log-fold-changes
}
\author{
Lukasz Jan Kielpinski
}

