% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutationEffect.R
\name{plotMutationEffect}
\alias{plotMutationEffect}
\title{Plot effects of substitutions}
\usage{
plotMutationEffect(stringSet, oneString, FoldChanges)
}
\arguments{
\item{oneString}{sequence of a reference kmer}

\item{FoldChanges}{logFC paired with stringSet}
}
\value{
Plot
}
\description{
Function plots what happens to log2-fold changes when one nucleotide 
of a reference sequence is substituted
}
\details{
#' @param stringSet Sequences of all kmers of a given length that have FoldChanges data
}
\author{
Lukasz Jan Kielpinski
}

