% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRunningCorrelation.R
\name{plotRunningCorrelation}
\alias{plotRunningCorrelation}
\title{Plot correlation of logFC with target knockdown at different sites within antisense oligonucleotides}
\usage{
plotRunningCorrelation(cleavabilityMatrix, observed, conf.level = 0.99, ...)
}
\arguments{
\item{cleavabilityMatrix}{Matrix with logFC scores calculated at different position of the oligonucleotides. 
Output of \code{\link{predictWithin}}}

\item{observed}{Experimental knockdown values}

\item{...}{Further parameters to \code{\link{plot}}}
}
\value{
Plot
}
\description{
Function plots correlation level with confidence intervals across set of equal-length oligos
It considers only bonds within the oligonucleotide.
}
\author{
Lukasz Jan Kielpinski
}
\seealso{
\code{\link{predictWithin}}
}

