% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCleavages.R
\name{predictCleavages}
\alias{predictCleavages}
\title{Predict cleavability of bonds within a given sequence by the HIV-1 RNase H}
\usage{
predictCleavages(rnaSequence, hbp)
}
\arguments{
\item{rnaSequence}{Sequence to be scored. Provide the RNA strand, but with T's (not U's)}

\item{hbp}{Heptamer Based Predictor. See details.}
}
\value{
Vector with cleavability scores (log2-fold-changes)
}
\description{
Function uses Heptamer Based Predictor ()
to assess the cleavability of a longer sequence by HIV-1 RNase H
}
\details{
Heptamer Based Predictor (hbp) is a list consisting of three vectors:

"kmers_seq" - vector of sequences of heptamers. Each kmer has to be represented.

"where_cut" - vector of cut sites (after which nt heptamer is cut). Calculated with \code{\link{findCleavageSite}}

"logFC_cut" - maximum observed extent of the cleavage of a given heptamer.

This method will report the best cleavage directed to each bond.
NA indicates that no heptamers direct the cleavage towards that bond.
}
\author{
Lukasz Jan Kielpinski
}

