% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictFC.R
\name{predictFC}
\alias{predictFC}
\title{Predict log2 fold changes based on a kmer-based-PWM}
\usage{
predictFC(charVector, scores)
}
\arguments{
\item{charVector}{Sequences to be scored. All need to be of the same length. 
The length has to match the length of scores [number of columns + kmer length - 1]}

\item{scores}{Matrix with scores; Subset of the output of the \code{\link{enrichmentMatrix}}}
}
\value{
predicted_scores
}
\description{
Function uses position-specific-kmer scoring matrix to score input sequences.
Input sequences must be of the same length and must be fully covered by the scoring matrix.
For single nucleotide model it is a sum of log2-fold-changes of those 
nucleotides from the scoring matrix that match the observed nucleotides. 
For longer models, kmers overlap, hence only a fraction of each kmers associated 
log2-fold-change is added to the prediction according to implemented weighing mechanism.
}
\author{
Lukasz Jan Kielpinski
}
\seealso{
For scoring longer sequence use: \code{\link{scoreSequence}}
}

