% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictWithin.R
\name{predictWithin}
\alias{predictWithin}
\title{Predict cleavability of a set of antisene oligonucleotide target sites}
\usage{
predictWithin(oligoSequences, scores, cleavageSiteInScores)
}
\arguments{
\item{oligoSequences}{Sequences of the antisense oligonucleotides}

\item{scores}{Matrix with scores; Subset of the output of the \code{\link{enrichmentMatrix}}}

\item{cleavageSiteInScores}{After which nucleotide of the scoring sequence the cleavage occurs. 
See details of \code{\link{scoreSequence}}}
}
\value{
cleavabilityMatrix Matrix with cleavability scores at different position
}
\description{
Function calculates the cleavability of oligo's target sites 
score using PWM of a set of oligonucleotides
}
\author{
Lukasz Jan Kielpinski
}

