% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSequence.R
\name{scoreSequence}
\alias{scoreSequence}
\title{Score sequence with PWM based scores}
\usage{
scoreSequence(rnaSequence, scores, cleavageSiteInScores)
}
\arguments{
\item{rnaSequence}{Sequence to be scored. Provide the RNA strand, but with T (not U)}

\item{scores}{Matrix with scores; Subset of the output of the \code{\link{enrichmentMatrix}}}

\item{cleavageSiteInScores}{After which nucleotide of the scoring sequence the cleavage occurs. See details}
}
\value{
Vector with logFC scores for each bond of the input sequence (including after last nt)
}
\description{
Function to score a single sequence for cleavability
given kmer-based PWM of RNase H sequence preferences
}
\details{
Assume that the scoring matrix (scores) has length i, and the atomic unit (kmer) length is j. 
Then the scoring sequence has the length i + j - 1. 
cleavageSiteInScores describes after which nucleotide within the scoring matrix the cleavage is expected to occur.
}
\author{
Lukasz Jan Kielpinski
}

