% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqLogoFromChars.R
\name{seqLogoFromChars}
\alias{seqLogoFromChars}
\title{Plot sequence logo from vector of characters}
\usage{
seqLogoFromChars(charVector, pwm_out = FALSE, plot = T, na.rm = T, ...)
}
\arguments{
\item{charVector}{Vector of equally-long characters; only ACGT allowed}

\item{pwm_out}{Should PWM be returned? (default = FALSE)}

\item{plot}{Should sequence logo be plotted? (default = TRUE)}

\item{na.rm}{Remove NA values prior to plotting (default = TRUE)}
}
\value{
pwmMatrix
}
\description{
Function to plots sequence logo using seqLogo package, 
but using vector of characters (of the same lenth) as input
}
\author{
Lukasz Jan Kielpinski
}

