% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBins.R
\name{summarizeBins}
\alias{summarizeBins}
\title{Summarize Bins}
\usage{
summarizeBins(referenceSet, sampledList, interval_starts, interval_stops)
}
\arguments{
\item{referenceSet}{Vector with numbers in the reference set}

\item{sampledList}{List of vectors with numbers in the background (e.g. sampled) sets}

\item{interval_starts}{Starts of the intervals [included]}

\item{interval_stops}{Stops of the intervals [included]}
}
\value{
List with (a) counts in intervals for the reference and (b) for a background and containing names of bins as column names
}
\description{
Function counts occurences of numbers within intervals of interest.
It does so for a reference set and for a list of sampled sets, 
creating a list suitable for \code{\link{makeViolinPlot}}
}
\author{
Lukasz Jan Kielpinski
}

